BAC-7362: add ImageDetailModal component and integrate image detail functionality
